Refactor CLI input helpers
